fix compiler waring (regarding size_t and 64bit architecture)
git-svn-id: svn://svn.minisip.org/minisip/trunk@2194 63dae60f-c5d3-0310-92a3-fcbf00261c72

+++ libminisip/source/conference/ConferenceControl.cxx
@@ -52,7 +52,7 @@ ConferenceControl::ConferenceControl(string configUri, string cid, bool islocal)
 	
 	confId=cid;
 	
-	uint32_t i = (uint32_t)configUri.find("@");
+	size_t i = (uint32_t)configUri.find("@");
 	assert(i!=string::npos);
 	myUri=configUri.substr(0,i);
 	myDomain=trim(configUri.substr(i));
